+++ src/main_window.rs
@@ -1,40 +1,35 @@
 use std::num::NonZero;
 use std::sync::Arc;
 
-use muda::ContextMenu;
+use muda::{ContextMenu, Menu, MenuItem};
 use raw_window_handle::{RawWindowHandle, Win32WindowHandle};
+use windows::core::{w, PCWSTR};
 use windows::Win32::Foundation::*;
-use windows::Win32::Graphics::Gdi::MapWindowPoints;
 use windows::Win32::UI::HiDpi::GetDpiForWindow;
-use windows::Win32::UI::Input::KeyboardAndMouse::*;
 use windows::Win32::UI::WindowsAndMessaging::*;
 use windows::UI::ViewManagement::{UIColorType, UISettings};
 use winit::dpi::PhysicalSize;
-use winit::event::WindowEvent;
-use winit::event_loop::ActiveEventLoop;
+use winit::event_loop::{ActiveEventLoop, EventLoopProxy};
 use winit::platform::windows::WindowAttributesExtWindows;
 use winit::window::{Window, WindowAttributes};
 
 use crate::app::{App, AppMessage};
 use crate::egui_glue::{EguiView, EguiWindow};
+use crate::host::create_host;
 use crate::komorebi::listen_for_workspaces;
 use crate::widgets::WorkspaceButton;
+use crate::window_registry_info::WindowRegistryInfo;
 
 impl App {
     pub fn create_main_window(&mut self, event_loop: &ActiveEventLoop) -> anyhow::Result<()> {
         tracing::info!("Creating main window");
 
-        let host = unsafe { crate::host::create_host(self.taskbar_hwnd, self.proxy.clone()) }?;
+        let window_info = WindowRegistryInfo::load()?;
+
+        let host = unsafe { create_host(self.taskbar_hwnd, self.proxy.clone(), &window_info) }?;
 
         let mut attrs = WindowAttributes::default();
-
-        // get host width/height
-        let mut rect = RECT::default();
-        unsafe { GetClientRect(host, &mut rect) }?;
-        let width = rect.right - rect.left;
-        let heigth = rect.bottom - rect.top;
-
-        attrs = attrs.with_inner_size(PhysicalSize::new(width, heigth));
+        attrs = attrs.with_inner_size(PhysicalSize::new(window_info.width, window_info.height));
 
         let parent = unsafe { NonZero::new_unchecked(host.0 as _) };
         let parent = Win32WindowHandle::new(parent);
@@ -58,7 +53,13 @@ impl App {
         let window = event_loop.create_window(attrs)?;
         let window = Arc::new(window);
 
-        let state = MainWindowView::new(window.clone(), host, self.taskbar_hwnd)?;
+        let state = MainWindowView::new(
+            window.clone(),
+            host,
+            self.taskbar_hwnd,
+            self.proxy.clone(),
+            window_info,
+        )?;
 
         let proxy = self.proxy.clone();
 
@@ -72,44 +73,40 @@ impl App {
     }
 }
 
-#[derive(Clone, Copy, PartialEq, PartialOrd, Debug)]
-enum DraggingState {
-    /// Not in drag mode
-    None,
-    /// In drag mode but user hasn't started dragging yet
-    Started,
-    /// In drag mode and actually dragging
-    Dragging,
-}
-
 pub struct MainWindowView {
     window: Arc<Window>,
     host: HWND,
     taskbar_hwnd: HWND,
-    size: (i32, i32),
-    workspaces: Vec<crate::komorebi::Workspace>,
+    proxy: EventLoopProxy<AppMessage>,
     context_menu: muda::Menu,
-    dragging_state: DraggingState,
+    workspaces: Vec<crate::komorebi::Workspace>,
     accent_light2_color: Option<egui::Color32>,
     accent_color: Option<egui::Color32>,
     forgreound_color: Option<egui::Color32>,
+    window_info: WindowRegistryInfo,
 }
 
 impl MainWindowView {
-    fn new(window: Arc<Window>, host: HWND, taskbar_hwnd: HWND) -> anyhow::Result<Self> {
+    fn new(
+        window: Arc<Window>,
+        host: HWND,
+        taskbar_hwnd: HWND,
+        proxy: EventLoopProxy<AppMessage>,
+        window_info: WindowRegistryInfo,
+    ) -> anyhow::Result<Self> {
         let workspaces = crate::komorebi::read_workspaces().unwrap_or_default();
 
         let mut view = Self {
             window,
             host,
+            proxy,
             taskbar_hwnd,
-            size: (0, 0),
             workspaces,
             context_menu: Self::create_context_menu()?,
-            dragging_state: DraggingState::None,
             accent_color: None,
             accent_light2_color: None,
             forgreound_color: None,
+            window_info,
         };
 
         if let Err(e) = view.update_system_colors() {
@@ -119,17 +116,15 @@ impl MainWindowView {
         Ok(view)
     }
 
-    /// The ID of the "Move" menu item.
-    const M_MOVE_ID: &str = "move";
+    /// The ID of the "Move & Resize" menu item.
+    const M_MOVE_RESIZE_ID: &str = "main-window-move-resize";
     /// The ID of the "Quit" menu item.
-    const M_QUIT_ID: &str = "quit";
+    const M_QUIT_ID: &str = "main-window-quit";
 
     fn create_context_menu() -> anyhow::Result<muda::Menu> {
-        muda::Menu::with_items(&[
-            &muda::MenuItem::with_id(Self::M_MOVE_ID, "Move", true, None),
-            &muda::MenuItem::with_id(Self::M_QUIT_ID, "Quit", true, None),
-        ])
-        .map_err(Into::into)
+        let quit = MenuItem::with_id(Self::M_QUIT_ID, "Quit", true, None);
+        let move_resize = MenuItem::with_id(Self::M_MOVE_RESIZE_ID, "Move & Resize", true, None);
+        Menu::with_items(&[&move_resize, &quit]).map_err(Into::into)
     }
 
     fn show_context_menu(&self) {
@@ -157,22 +152,16 @@ impl MainWindowView {
         Ok(())
     }
 
-    fn host_client_rect(&self) -> anyhow::Result<RECT> {
-        let mut rect = RECT::default();
-        unsafe { GetClientRect(self.host, &mut rect) }?;
-        Ok(rect)
-    }
-
     fn host_window_rect(&self) -> anyhow::Result<RECT> {
         let mut rect = RECT::default();
         unsafe { GetWindowRect(self.host, &mut rect) }?;
         Ok(rect)
     }
 
-    fn taskbar_client_rect(&self) -> anyhow::Result<RECT> {
+    fn taskbar_height(&self) -> anyhow::Result<i32> {
         let mut rect = RECT::default();
         unsafe { GetClientRect(self.taskbar_hwnd, &mut rect) }?;
-        Ok(rect)
+        Ok(rect.bottom - rect.top)
     }
 
     const WORKSPACES_MARGIN: egui::Margin = egui::Margin::same(1);
@@ -181,65 +170,48 @@ impl MainWindowView {
         let rect = rect + Self::WORKSPACES_MARGIN;
         let rect = rect * ppp;
 
-        let width = rect.width() as i32;
-        // height always matches the taskbar height
-        let taskbar_rect = self.taskbar_client_rect()?;
-        let height = taskbar_rect.bottom - taskbar_rect.top;
+        let height = if self.window_info.auto_height {
+            self.taskbar_height()?
+        } else {
+            self.window_info.height
+        };
 
-        let (curr_width, curr_height) = self.size;
+        let width = if self.window_info.auto_width {
+            rect.width() as i32
+        } else {
+            self.window_info.width
+        };
+
+        let curr_width = self.window_info.width;
+        let curr_height = self.window_info.height;
 
         if curr_width != width || curr_height != height {
-            self.size = (width, height);
+            self.window_info.width = width;
+            self.window_info.height = height;
 
             tracing::debug!("Resizing host to match content rect");
 
+            self.window_info.save()?;
             unsafe { SetWindowPos(self.host, None, 0, 0, width, height, SWP_NOMOVE) }?;
         }
-        Ok(())
-    }
-
-    fn start_host_dragging(&mut self) -> anyhow::Result<()> {
-        // start dragging mode
-        self.dragging_state = DraggingState::Started;
-
-        // get host width and height
-        let rect = self.host_client_rect()?;
-        let width = rect.right - rect.left;
-        let height = rect.bottom - rect.top;
-
-        // set the cursor position to the center of the host window
-        let x = rect.left + width / 2;
-        let y = rect.top + height / 2;
-        let points = &mut [POINT { x, y }];
-        unsafe { MapWindowPoints(Some(self.host), None, points) };
-        unsafe { SetCursorPos(points[0].x, points[0].y)? };
 
         Ok(())
     }
 
-    fn drag_host_window(&mut self) -> anyhow::Result<()> {
-        // get current cursor pos
-        let mut pos = POINT::default();
-        unsafe { GetCursorPos(&mut pos) }?;
+    pub const IN_RESIZE_PROP: PCWSTR = w!("komorebi::in_resize");
 
-        let points = POINTS {
-            x: pos.x as i16,
-            y: pos.y as i16,
-        };
+    fn start_host_dragging(&self) -> anyhow::Result<()> {
+        let host = self.host.0 as isize;
+        let info = self.window_info;
+        let message = AppMessage::CreateResizeWindow { host, info };
+        self.proxy.send_event(message)?;
+        unsafe { SetPropW(self.host, Self::IN_RESIZE_PROP, Some(HANDLE(1 as _))) }?;
+        Ok(())
+    }
 
-        unsafe {
-            // release cursor capture
-            ReleaseCapture()?;
-
-            // simulate left click on the host window invisible title bar
-            PostMessageW(
-                Some(self.host),
-                WM_NCLBUTTONDOWN,
-                WPARAM(HTCAPTION as _),
-                LPARAM(&points as *const _ as _),
-            )?;
-        }
-
+    fn update_window_info(&mut self, info: &WindowRegistryInfo) -> anyhow::Result<()> {
+        self.window_info = *info;
+        unsafe { RemovePropW(self.host, Self::IN_RESIZE_PROP) }?;
         Ok(())
     }
 
@@ -263,7 +235,7 @@ impl MainWindowView {
             .map(|r| {
                 let current_monitor = self.window.current_monitor();
                 let y = current_monitor.map(|m| m.position().y).unwrap_or(0);
-                r.top == y
+                r.top <= y
             })
             .unwrap_or(false)
     }
@@ -284,20 +256,6 @@ impl MainWindowView {
     }
 
     fn workspaces_row(&mut self, ui: &mut egui::Ui) -> egui::Response {
-        // if dragging
-        if self.dragging_state == DraggingState::Started {
-            // change the cursor
-            ui.ctx().set_cursor_icon(egui::CursorIcon::ResizeColumn);
-
-            // start dragging on the first left click
-            if ui.input(|i| i.pointer.button_down(egui::PointerButton::Primary)) {
-                self.dragging_state = DraggingState::Dragging;
-                if let Err(e) = self.drag_host_window() {
-                    tracing::error!("Failed to start host darggign: {e}");
-                }
-            }
-        }
-
         // show context menu on right click
         if ui.input(|i| i.pointer.button_pressed(egui::PointerButton::Secondary)) {
             self.show_context_menu();
@@ -314,7 +272,7 @@ impl MainWindowView {
                         .text_color_opt(self.forgreound_color)
                         .line_on_top(self.is_taskbar_on_top());
 
-                    if ui.add(btn).clicked() && self.dragging_state == DraggingState::None {
+                    if ui.add(btn).clicked() {
                         crate::komorebi::change_workspace(workspace.idx);
                     }
                 }
@@ -324,14 +282,8 @@ impl MainWindowView {
     }
 
     fn transparent_panel(&self, ctx: &egui::Context) -> egui::CentralPanel {
-        let color = if self.dragging_state != DraggingState::None {
-            egui::Color32::from_rgba_unmultiplied(0, 0, 0, 125)
-        } else {
-            egui::Color32::TRANSPARENT
-        };
-
         let visuals = egui::Visuals {
-            panel_fill: color,
+            panel_fill: egui::Color32::TRANSPARENT,
             ..egui::Visuals::dark()
         };
         ctx.set_visuals(visuals);
@@ -343,24 +295,6 @@ impl MainWindowView {
 }
 
 impl EguiView for MainWindowView {
-    fn handle_window_event(
-        &mut self,
-        _ctx: &egui::Context,
-        _event_loop: &ActiveEventLoop,
-        event: winit::event::WindowEvent,
-    ) -> anyhow::Result<()> {
-        // exit dragging mode on cursor enter if we are in dragging mode
-        // this is done here so the dragging mode background
-        // doesn't change until dragging is over
-        if self.dragging_state == DraggingState::Dragging
-            && matches!(event, WindowEvent::CursorEntered { .. })
-        {
-            self.dragging_state = DraggingState::None;
-        }
-
-        Ok(())
-    }
-
     fn handle_app_message(
         &mut self,
         ctx: &egui::Context,
@@ -369,9 +303,13 @@ impl EguiView for MainWindowView {
     ) -> anyhow::Result<()> {
         match message {
             AppMessage::UpdateWorkspaces(workspaces) => self.workspaces = workspaces.clone(),
-            AppMessage::MenuEvent(e) if e.id() == Self::M_MOVE_ID => self.start_host_dragging()?,
+            AppMessage::MenuEvent(e) if e.id() == Self::M_MOVE_RESIZE_ID => {
+                self.start_host_dragging()?
+            }
+            AppMessage::StartMoveResize => self.start_host_dragging()?,
             AppMessage::MenuEvent(e) if e.id() == Self::M_QUIT_ID => self.close_host()?,
             AppMessage::SystemSettingsChanged => self.update_system_colors()?,
+            AppMessage::NotifyWindowInfoChanges(info) => self.update_window_info(info)?,
             AppMessage::DpiChanged => {
                 let dpi = unsafe { GetDpiForWindow(self.host) } as f32;
                 let ppp = dpi / USER_DEFAULT_SCREEN_DPI as f32;